fix: rename the package
